+++ src/navigation-bar.tsx
@@ -3,6 +3,7 @@ import {CSSTransition} from 'react-transition-group';
 import classnames from 'classnames';
 import ResponsiveLayout from './responsive-layout';
 import Inline from './inline';
+import Box from './box';
 import Touchable from './touchable';
 import {Text2, Text3} from './text';
 import {useScreenSize, useTheme, useAriaId} from './hooks';
@@ -194,6 +195,10 @@ const useStyles = createUseStyles((theme) => {
             right: 0,
             zIndex: NAVBAR_ZINDEX,
         },
+        notFixedPadding: {
+            width: '100%',
+            padding: ({paddingX}) => `0 ${paddingX}px`,
+        },
         navbar: {
             width: '100%',
             display: 'flex',
@@ -202,11 +207,12 @@ const useStyles = createUseStyles((theme) => {
                 isInverse ? theme.colors.navigationBarBackground : theme.colors.background,
             height: DESKTOP_NAVBAR_HEIGHT,
             padding: '16px 0',
-            transition: 'border 300ms',
-            borderBottom: `1px solid ${theme.colors.divider}`,
+            borderBottomStyle: 'solid',
+            borderBottomWidth: ({withBorder}) => (withBorder ? 1 : 0),
             borderColor: ({isInverse}) =>
                 isInverse && !theme.isDarkMode ? 'transparent' : theme.colors.divider,
             [theme.mq.tabletOrSmaller]: {
+                transition: 'border-color 300ms',
                 borderColor: ({isMenuOpen, isInverse}) =>
                     isMenuOpen || (isInverse && !theme.isDarkMode) ? 'transparent' : theme.colors.divider,
                 height: MOBILE_NAVBAR_HEIGHT,
@@ -221,11 +227,7 @@ const useStyles = createUseStyles((theme) => {
             borderBottom: `2px solid transparent`,
             transition: 'border-color 300ms ease-in-out',
 
-            // Only apply hover effect to user agents using fine pointer devices (a mouse, for example)
-            // Also enabled for (pointer: none) for acceptance tests, where (pointer: fine) doesn't match.
-            // WARNING: you may be tempted to use @media (hover: hover) instead, but that doesn't work as expected in some android browsers.
-            // See: https://hover-pointer-media-query.glitch.me/ and https://github.com/mui-org/material-ui/issues/15736
-            '@media (pointer: fine), (pointer: none)': {
+            [theme.mq.supportsHover]: {
                 '&:hover span': {
                     color: ({isInverse}) =>
                         isInverse ? theme.colors.textSecondaryInverse : theme.colors.textSecondary,
@@ -266,6 +268,14 @@ const useStyles = createUseStyles((theme) => {
         burgerMenuExitActive: {
             transform: 'translate(-100vw)',
         },
+        iconButton: {
+            color: ({isInverse}) => (isInverse ? theme.colors.inverse : theme.colors.neutralHigh),
+            [theme.mq.supportsHover]: {
+                '&:hover': {
+                    color: ({isInverse}) => (isInverse ? theme.colors.inverse : theme.colors.neutralMedium),
+                },
+            },
+        },
     };
 });
 
@@ -298,7 +308,7 @@ export const MainNavigationBar: React.FC<MainNavigationBarProps> = ({
     const [isMenuOpen, setIsMenuOpen] = React.useState(false);
     const [menuTransitionState, setMenuTransitionState] = React.useState<MenuTransitionState>('closed');
     const menuId = useAriaId();
-    const classes = useStyles({isMenuOpen, isInverse, menuTransitionState});
+    const classes = useStyles({isMenuOpen, isInverse, menuTransitionState, withBorder: true});
     const {isTabletOrSmaller} = useScreenSize();
     const setModalState = useSetModalState();
 
@@ -432,14 +442,26 @@ export const MainNavigationBar: React.FC<MainNavigationBarProps> = ({
     );
 };
 
-type NavigationBarProps = {
+interface NavigationBarCommonProps {
     isInverse?: boolean;
     onBack?: () => void;
     title?: string;
     right?: React.ReactElement;
-    topFixed?: boolean;
+    withBorder?: boolean;
     children?: undefined;
-};
+}
+
+interface NavigationBarTopFixedProps extends NavigationBarCommonProps {
+    topFixed?: true;
+    paddingX?: undefined;
+}
+
+interface NavigationBarNotFixedProps extends NavigationBarCommonProps {
+    topFixed: false;
+    paddingX?: number;
+}
+
+type NavigationBarProps = NavigationBarTopFixedProps | NavigationBarNotFixedProps;
 
 export const NavigationBar: React.FC<NavigationBarProps> = ({
     onBack,
@@ -447,25 +469,38 @@ export const NavigationBar: React.FC<NavigationBarProps> = ({
     right,
     isInverse = false,
     topFixed = true,
+    paddingX,
+    withBorder = true,
 }) => {
-    const classes = useStyles({isInverse});
+    const {texts} = useTheme();
+    const classes = useStyles({isInverse, paddingX: paddingX ?? 0, withBorder});
     const content = (
         <Inline space="between" alignItems="center">
             <Inline space={24} alignItems="center">
                 {onBack && (
-                    <IconButton aria-label="back" onPress={onBack}>
-                        <IconChevronLeftRegular />
+                    <IconButton
+                        aria-label={texts.backNavigationBar}
+                        onPress={onBack}
+                        className={classes.iconButton}
+                    >
+                        <IconChevronLeftRegular color="currentColor" />
                     </IconButton>
                 )}
-                <Text3 regular>{title}</Text3>
+                <Text3 regular truncate>
+                    {title}
+                </Text3>
             </Inline>
-            {right}
+            <Box paddingLeft={24}>{right}</Box>
         </Inline>
     );
     return (
         <ThemeVariant isInverse={isInverse}>
             <header className={classnames(classes.navbar, {[classes.topFixed]: topFixed})}>
-                {topFixed ? <ResponsiveLayout>{content}</ResponsiveLayout> : content}
+                {topFixed ? (
+                    <ResponsiveLayout>{content}</ResponsiveLayout>
+                ) : (
+                    <div className={classes.notFixedPadding}>{content}</div>
+                )}
             </header>
             {topFixed && <div className={classes.spacer} />}
         </ThemeVariant>
@@ -486,7 +521,7 @@ export const FunnelNavigationBar: React.FC<FunnelNavigationBarProps> = ({
     isInverse = false,
     topFixed = true,
 }) => {
-    const classes = useStyles({isInverse});
+    const classes = useStyles({isInverse, withBorder: true});
     return (
         <ThemeVariant isInverse={isInverse}>
             <header className={classnames(classes.navbar, {[classes.topFixed]: topFixed})}>
@@ -512,7 +547,11 @@ const useNavigationBarActionGroupStyles = createUseStyles(() => ({
     },
 }));
 
-export const NavigationBarActionGroup: React.FC = ({children}) => {
+type NavigationBarActionGroupProps = {
+    children: React.ReactNode;
+};
+
+export const NavigationBarActionGroup: React.FC<NavigationBarActionGroupProps> = ({children}) => {
     const classes = useNavigationBarActionGroupStyles();
     return (
         <div className={classes.lineHeightFix}>
@@ -532,11 +571,7 @@ const useNavigationBarActionStyles = createUseStyles((theme) => ({
             color: ({isInverse}) => (isInverse ? theme.colors.inverse : theme.colors.neutralHigh),
         },
 
-        // Only apply hover effect to user agents using fine pointer devices (a mouse, for example)
-        // Also enabled for (pointer: none) for acceptance tests, where (pointer: fine) doesn't match.
-        // WARNING: you may be tempted to use @media (hover: hover) instead, but that doesn't work as expected in some android browsers.
-        // See: https://hover-pointer-media-query.glitch.me/ and https://github.com/mui-org/material-ui/issues/15736
-        '@media (pointer: fine), (pointer: none)': {
+        [theme.mq.supportsHover]: {
             '&:hover span': {
                 color: ({isInverse}) =>
                     isInverse ? theme.colors.textSecondaryInverse : theme.colors.textSecondary,