+++ src/password-field.tsx
@@ -13,9 +13,9 @@ export interface PasswordFieldProps extends CommonFormFieldProps {
     onChangeValue?: (value: string, rawValue: string) => void;
 }
 
-const usePasswordAdornmentStyles = createUseStyles(() => ({
+const usePasswordAdornmentStyles = createUseStyles((theme) => ({
     shadow: {
-        ['@media (hover: hover)']: {
+        [theme.mq.supportsHover]: {
             '&:hover': {
                 backgroundColor: 'rgba(0, 0, 0, 0.08)',
             },
@@ -69,6 +69,7 @@ const PasswordField: React.FC<PasswordFieldProps> = ({
     ...rest
 }) => {
     const [isVisible, setIsVisible] = React.useState(false);
+    const caretPositionRef = React.useRef<number>(0);
     const inputRef = React.useRef<HTMLInputElement>(null);
 
     const processValue = (value: string) => value;
@@ -76,15 +77,20 @@ const PasswordField: React.FC<PasswordFieldProps> = ({
     const focus = () => {
         const input = inputRef.current;
         if (input) {
+            if (input.selectionStart !== null) {
+                caretPositionRef.current = input.selectionStart;
+            }
             input.focus();
-            // neeeded to place the caret at the end
-            setTimeout(() => {
-                const v = input.value;
-                input.value = '';
-                input.value = v;
-            }, 0);
         }
     };
+
+    React.useEffect(() => {
+        const input = inputRef.current;
+        if (input) {
+            input.selectionStart = caretPositionRef.current;
+            input.selectionEnd = caretPositionRef.current;
+        }
+    }, [isVisible, caretPositionRef, inputRef]);
 
     const fieldProps = useFieldProps({
         name,