+++ src/skins/movistar.tsx
@@ -160,7 +160,7 @@ export const getMovistarSkin: GetSkin = (variant) => {
             textPrimaryInverse: palette.white,
             textSecondary: palette.grey5,
             textSecondaryInverse: palette.white,
-            textDisabled: palette.grey3,
+            textDisabled: palette.grey4,
             textAmount: palette.movistarBlue,
 
             // STATES