+++ src/skins/telefonica.tsx
@@ -155,7 +155,7 @@ export const getTelefonicaSkin: GetSkin = () => {
             textPrimaryInverse: palette.white,
             textSecondary: palette.grey5,
             textSecondaryInverse: palette.white,
-            textDisabled: palette.grey3,
+            textDisabled: palette.grey4,
             textAmount: palette.telefonicaBlue,
 
             // STATES