+++ src/text-field-base.tsx
@@ -144,7 +144,8 @@ const commonInputStyles = (theme: Theme) => ({
         },
     },
     '&:disabled': {
-        color: theme.colors.border,
+        color: theme.colors.textDisabled,
+        cursor: 'not-allowed',
     },
     boxShadow: 'none', // reset FF red shadow styles for required inputs
 });
@@ -228,6 +229,7 @@ const useStyles = createUseStyles((theme) => ({
         paddingRight: 16,
         display: 'flex',
         alignItems: 'center',
+        opacity: ({disabled}) => (disabled ? 0.3 : 1),
     },
     startIcon: {
         pointerEvents: 'none', // passthrough click events to the input
@@ -237,6 +239,7 @@ const useStyles = createUseStyles((theme) => ({
         alignItems: 'center',
         height: '100%',
         position: 'absolute',
+        opacity: ({disabled}) => (disabled ? 0.3 : 1),
     },
     prefix: {
         paddingTop: ({hasLabel}) => (hasLabel ? 25 : 16),
@@ -245,7 +248,7 @@ const useStyles = createUseStyles((theme) => ({
         paddingRight: 16,
         display: 'flex',
         alignItems: 'center',
-        color: theme.colors.textSecondary,
+        color: ({disabled}) => (disabled ? theme.colors.textDisabled : theme.colors.textSecondary),
         opacity: ({inputState}) => (inputState === 'default' ? 0 : 1),
         transition: 'opacity 150ms cubic-bezier(0.0, 0, 0.2, 1) 0ms',
     },
@@ -307,6 +310,7 @@ const TextFieldBaseComponent = React.forwardRef<any, TextFieldBaseProps>(
             prefix,
             multiline,
             type: rest.type,
+            disabled: rest.disabled,
         });
 
         React.useEffect(() => {
@@ -348,8 +352,19 @@ const TextFieldBaseComponent = React.forwardRef<any, TextFieldBaseProps>(
             ...inputProps,
         };
 
+        let labelStyle = {};
+        const isShrinked = shrinkLabel || inputState === 'focused' || inputState === 'filled';
+        if (startIcon) {
+            labelStyle = isShrinked
+                ? {left: 48, right: 0, width: 'auto'}
+                : {left: 48, right: 12, width: 'auto'};
+        } else if (endIcon && !isShrinked) {
+            labelStyle = {paddingRight: 36};
+        }
+
         return (
             <FieldContainer
+                disabled={rest.disabled}
                 helperText={
                     <HelperText
                         error={error}
@@ -396,7 +411,7 @@ const TextFieldBaseComponent = React.forwardRef<any, TextFieldBaseProps>(
                 })}
                 {label && (
                     <Label
-                        style={startIcon ? {marginLeft: 48, left: 0} : {}}
+                        style={labelStyle}
                         error={error}
                         forId={id}
                         inputState={inputState}